requirements.txt
astroid==1.6.4
colorama==0.3.9
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
nltk==3.3
numpy==1.14.3
paho-mqtt==1.3.1
pandas==0.23.0
pylint==1.9.1
pymongo==3.7.0
python-dateutil==2.7.3
pytz==2018.4
scikit-learn==0.19.1
scipy==1.1.0
six==1.11.0
wrapt==1.10.11
